requirements.txt
Flask-JWT
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy